+++ frontend/src/components/title/TitleGrid.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import Grid from './styled-components/Grid';
-import Title from './Title';
-import Header from './Header';
+import Title from './components/title/Title';
+import Header from './components/header/Header';
 
 const TitleGrid = (props) => (
-	<Grid>
-		<Title />
-		<Header showIcons={props.showIcons} />
-	</Grid>
+    <Grid>
+        <Title />
+        <Header showIcons={props.showIcons} />
+    </Grid>
 );
 
 export default TitleGrid;
