+++ frontend/src/components/title/components/details/Details.jsx
