+++ frontend/src/components/title/components/details/styled-components/DetailsContainer.jsx
