+++ frontend/src/components/title/components/details/styled-components/GreyText.jsx
