+++ frontend/src/components/title/components/details/styled-components/RatingText.jsx
