+++ frontend/src/components/title/components/icons/Icons.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import Share from '../../buttons/Share';
+import Save from '../../buttons/Save';
+
+const Icons = () => (
+    <React.Fragment>
+        <Share />
+        <Save />
+    </React.Fragment>
+);
+
+export default Icons;
