+++ package.json
@@ -53,11 +53,14 @@
         "babel-loader": "^8.2.2",
         "cors": "^2.8.5",
         "css-loader": "^5.0.1",
+        "eslint-plugin-jest-dom": "^3.6.5",
+        "eslint-plugin-testing-library": "^3.10.1",
         "file-loader": "^6.2.0",
         "html-webpack-plugin": "^5.0.0-alpha.3",
         "jest": "^26.6.3",
         "jest-svg-transformer": "^1.0.0",
         "mini-css-extract-plugin": "^1.3.3",
+        "msw": "^0.25.0",
         "node-fetch": "^2.6.1",
         "request": "^2.88.2",
         "source-map-loader": "^2.0.0",
